+++ markets/spot-market/contracts/storage/FeeConfiguration.sol
@@ -354,7 +354,7 @@ library FeeConfiguration {
             uint previousUsdBalance = spotMarketFactory.usdToken.balanceOf(address(this));
 
             spotMarketFactory.usdToken.approve(address(feeCollector), totalFeesUint);
-            feeCollector.collectFees(marketId, totalFeesUint, transactor, uint8(transactionType)); // TODO dont think we have safe casts for uint8
+            feeCollector.collectFees(marketId, totalFeesUint, transactor, uint(transactionType));
 
             uint currentUsdBalance = spotMarketFactory.usdToken.balanceOf(address(this));
             collectedFees = previousUsdBalance - currentUsdBalance;
@@ -385,7 +385,7 @@ library FeeConfiguration {
                 marketId,
                 totalFeesUint,
                 transactor,
-                uint8(transactionType)
+                uint(transactionType)
             );
         }
     }
